+++ src/gaplessgrid.c
